# fastrand: optimize Buf